+++ bzt/modules/blazemeter/cloud_provisioning.py
@@ -46,7 +46,6 @@ from bzt.modules.blazemeter.net_utils import get_with_retry
 from bzt.modules.blazemeter.blazemeter_reporter import BlazeMeterUploader
 from bzt.modules.blazemeter.cloud_test import FUNC_API_TEST_TYPE, FUNC_GUI_TEST_TYPE, TAURUS_TEST_TYPE
 from bzt.modules.blazemeter.project_finder import ProjectFinder
-from bzt.modules.blazemeter.stubs import ServiceStubCaptureHAR
 from bzt.modules.blazemeter.cloud_const import DEDICATED_IPS
 
 CLOUD_CONFIG_BLACK_LIST = {
@@ -820,3 +819,15 @@ class CloudProvWidget(Pile, PrioritizedWidget):
         txt = self.test.get_test_status_text()
         if txt:
             self.text.set_text(txt)
+
+
+class ServiceStubScreenshoter(Service):
+    def startup(self):
+        if not isinstance(self.engine.provisioning, CloudProvisioning):
+            self.log.warning("Stub for service 'screenshoter', use cloud provisioning to have it working")
+
+
+class ServiceStubCaptureHAR(Service):
+    def startup(self):
+        if not isinstance(self.engine.provisioning, CloudProvisioning):
+            self.log.warning("Stub for service 'capturehar', use cloud provisioning to have it working")
