+++ tests/unit/modules/test_blazeMeterUploader.py
@@ -12,7 +12,7 @@ from bzt import TaurusException
 from bzt.bza import Master, Session
 from bzt.modules.aggregator import DataPoint, KPISet
 from bzt.modules.blazemeter import BlazeMeterUploader
-from bzt.modules.blazemeter import MonitoringBuffer
+from bzt.modules.blazemeter.blazemeter_reporter import MonitoringBuffer
 from bzt.utils import iteritems, viewvalues
 from tests.unit import BZTestCase, random_datapoint, RESOURCES_DIR, ROOT_LOGGER, EngineEmul, BZMock
